Moved Other User's Wishlists to a Bootstrap Carousel. This eliminates the need for the click/select on the select element, and makes it more clear who you are shopping for at a glance.
index-carousel is a file where the tab system has been adapted into a
carousel. It was not implemented because it creates a problem in the
nested carousel on other user's wishlists. I have mixed feelings about it
as it provides a more interactive experience, but can also be a little
jarring. It's here for historical purposes. It appears bootstrap/twitter
is not interested in implementing nested carousels (see
https://github.com/twitter/bootstrap/issues/4359)

Script.js and Style.css files account for changes listed above.