DB.class - Just removed a commented out old method.
index.php
- Fixed Manage image form button disabled/clickable/accessible on add item (Issue #11)
- Changed button from anchor link to button element to support disabled click event handling.
- Fixed userId assignment on page load.
- Added Manage user form & button in manage section
- Added more data to manage section and admin visibility.

liveTest.php
This file is for making calls to methods in classes without having to go through the front-end/ajax interface. Significant work has been done of this page to make it more user-friendly and improve support for additional classes/methods. It now automatically parses files based on inline documentation and builds not only a select set of methods to call, but also a variable set to pass to the system.

script.js
- Removed getItemImages method as it's redundant.
- Added populateMangeUserForm method for handling getting and displaying the manage user form.

user.class.php
- Added getUser and manageUser methods for managing users in the system.

wishlist.class.php
- Just documentation cleanup to match standard across the project.